requirements.txt
flask
flask-wtf
transformers
numpy
scipy
pillow
os
